% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lagmat}
\alias{lagmat}
\title{Lag matrix of a time series}
\usage{
lagmat(obs, k)
}
\arguments{
\item{obs}{a numeric vector of observations}

\item{k}{an integer indicating the number of lags.}
}
\value{
The return object is a matrix. Each row corresponds to k lags of the observation.
The dimension of the matrix is truncated data size x k.
}
\description{
This function computes a given number of lags for each observation of a time series.
}
