% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtd_wrapper_function.R
\name{mtd}
\alias{mtd}
\title{Plots a Modified Target Diagram (MTD)}
\usage{
mtd(
  pollutant,
  dates,
  reference_data,
  sensor_data,
  ubs_rm,
  ubss,
  unit_ref = "ug/m3",
  unit_sensor = "ug/m3",
  x_axis_label = NULL,
  y_axis_label = NULL,
  x_lim = NA,
  y_lim = NA,
  sensor_name = NULL
)
}
\arguments{
\item{pollutant}{Required. String giving pollutant (one of "CO","CO2","CO2","NO","NO2","O3","PM1","PM10","PM2.5")}

\item{dates}{Required. Vector of datetimes of the measurements}

\item{reference_data}{Required. Numeric vector of observations from reference}

\item{sensor_data}{Required. Numeric vector of observations from sensor}

\item{ubs_rm}{Required. Random standard uncertainty of the results of the reference method, given as a constant value for all reference values (in same units as measurements)}

\item{ubss}{Required. Between standard uncertainty of the sensor, given as a constant value for all yis sensor values (in same units as measurements)}

\item{unit_ref}{Optional. Character string prodiving the unit of the reference values (one of "ppm", "ppb", "ug/m3", "mg/m3"). Default is 'ug/m3'.}

\item{unit_sensor}{Optional. Character string prodiving the unit of the sensor values (one of "ppm", "ppb", "ug/m3", "mg/m3"). Default is 'ug/m3'.}

\item{x_axis_label, y_axis_label}{Optional. Character string. Labels of the x and y axes.}

\item{x_lim, y_lim}{Optional. Numeric vectors of two values min and max values. Limits of x and y axis, default values is NA. [DOES NOT WORK CURRENTLY]}

\item{sensor_name}{Optional. Character string. Name of the sensor to be written in front of the calibration equation. If NULL (default), do not print sensor name.}
}
\value{
A plot
}
\description{
This function creates a plot of a Modified Target Diagram according to Yatkin et al. (2022). Such diagrams are useful for comparing, for example,
a low-cost air quality sensor against reference instrumentation, particularly with regard to the Data Quality Objectives of the European air quality directive.
}
\examples{
data(pm25)
mtd(pollutant = 'PM2.5',
    dates = pm25$datetime,
    reference_data = pm25$reference_pm25,
    sensor_data = pm25$sensor_pm25,
    ubs_rm = 0.5,
    ubss = 5)
}
\references{
Yatkin, S., Gerboles, M., Borowiak, A., Davila, S., Spinelle, L., Bartonova, A., Dauge, F., Schneider, P., Van Poppel, M., Peters, J.,
Matheeussen, C., & Signorini, M. (2022). Modified Target Diagram to check compliance of low-cost sensors with the Data Quality Objectives of the
European air quality directive. Atmospheric Environment, 273, 118967. \href{https://doi.org/10.1016/j.atmosenv.2022.118967}{https://doi.org/10.1016/j.atmosenv.2022.118967}
}
