% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{predMTD}
\alias{predMTD}
\title{One-step ahead prediction}
\usage{
predMTD(res, family, lags, probs)
}
\arguments{
\item{res}{a list of MCMC results from the MTD model; an output from functions 
either \code{\link{tsMTD}} or \code{\link{tsRegMTD}}.}

\item{family}{a quoted keyword that specifies the stationary marginal distribution.
Supported keywords are: "gaussian", "poisson", "negative binomial".}

\item{lags}{a numeric matrix of the lags of the observation. Each row 
corresponds to an observation. The dimension of the matrix is 
truncated data size x mtdorder.}

\item{probs}{a numeric vector of probabilities with values in \eqn{[0, 1]}.}
}
\description{
This function generates one-step-ahead posterior predictive distribution 
given a fitted MTD model.
}
\references{
Zheng, X., Kottas, A., and Sansó, B. (2021),
"On construction and estimation of stationary mixture transition distribution models,"
\emph{arXiv:2010.12696}.
}
