% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{rqrMTD}
\alias{rqrMTD}
\title{Randomize quantile residual analysis for MTD models}
\usage{
rqrMTD(res, obs, family, covar = NULL)
}
\arguments{
\item{res}{a list of MCMC results from the MTD model; an output from functions either \code{\link{tsMTD}} or \code{\link{tsRegMTD}}.}

\item{obs}{a numeric vector of complete data (not truncated).}

\item{family}{a quoted keyword that specifies the stationary marginal distribution.
Supported keywords are: "gaussian", "poisson" and "negative binomial".}

\item{covar}{covariates. The default is NULL. The argument is specified when a regression MTD model is desired.}
}
\value{
The return object is a matrix. Each column corresponds to posterior samples of an iteration.
The dimension of the matrix is truncated data size x posterior sample size.
}
\description{
This function performs model checking for MTD models based on the computed randomized quantile residuals.
}
\references{
Dunn, P. K. and Smyth, G. K. (1996), 
“Randomized quantile residuals,”
\emph{Journal of Computational and Graphical Statistics}.

Zheng, X., Kottas, A., and Sansó, B. (2021),
"On construction and estimation of stationary mixture transition distribution models,"
\emph{arXiv:2010.12696}.
}
