% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize.numeric}
\alias{summarize.numeric}
\title{Obtain posterior estimates}
\usage{
\method{summarize}{numeric}(x, point_est = "mean", lower_prob = 0.025, upper_prob = 0.975, k = 2)
}
\arguments{
\item{x}{a numeric vector.}

\item{point_est}{a quoted keyword that specifies the point estimate.
Supported keywords are "mean", "median", "both".}

\item{lower_prob}{Lower probability.}

\item{upper_prob}{Upper probability.}

\item{k}{an integer indicating the number of decimal places to be used.}
}
\description{
Obtain posterior estimates
}
