% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize}
\alias{summarize}
\title{Obtain posterior estimates}
\usage{
summarize(x, point_est, lower_prob = 0.025, upper_prob = 0.975, k = 2)
}
\arguments{
\item{x}{an object for which "summarize" is desired.}

\item{point_est}{a quoted keyword that specifies the point estimate.
Supported keywords are "mean", "median", "both".}

\item{lower_prob}{Lower probability.}

\item{upper_prob}{Upper probability.}

\item{k}{an integer indicating the number of decimal places to be used.}
}
\description{
Obtain posterior estimates
}
