% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMTD.R
\name{tsMTD}
\alias{tsMTD}
\title{Fit a time series MTD model}
\usage{
tsMTD(obs, mtdorder, family, weight, prior, tuning, starting, mcmc_param)
}
\arguments{
\item{obs}{a vector of complete data (not truncated).}

\item{mtdorder}{the model order. Note mtdorder > 1.}

\item{family}{a quoted keyword that specifies the stationary marginal distribution.
Supported keywords are: "gaussian", "poisson", "negative binomial".}

\item{weight}{a quoted keyword that specifies the prior for the weights.
Supported keywords are: "dir", "sb", "cdp".}

\item{prior}{a list of priors for model parameters.}

\item{tuning}{a list of tuning parameters  for model parameters.}

\item{starting}{a list of starting values for model parameters.}

\item{mcmc_param}{a list of MCMC parameters that contains, 
niter: the number of iterations;
nburn: the number of burn-in samples;
nthin: the number of steps to obtain thinned samples.}
}
\description{
This function fits an MTD model with a pre-specified stationary marginal distribution.
}
\references{
Zheng, X., Kottas, A., and Sansó, B. (2021),
"On construction and estimation of stationary mixture transition distribution models,"
\emph{arXiv:2010.12696}.
}
