% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMTD.R
\name{tsRegMTD}
\alias{tsRegMTD}
\title{Fit a time series regression MTD model}
\usage{
tsRegMTD(
  formula,
  mtdorder,
  family,
  weight,
  prior,
  tuning,
  starting,
  mcmc_param
)
}
\arguments{
\item{formula}{a symbolic description of the regression model. 
Depending on the marginal distribution of the MTD, the regression
model can be additive or multiplicative.}

\item{mtdorder}{the model order. Note mtdorder > 1.}

\item{family}{a quoted keyword that specifies the stationary marginal distribution.
Supported keyword is "lomax".}

\item{weight}{a quoted keyword that specifies the prior for the weights.
Supported keywords are: "dir", "sb", "cdp".}

\item{prior}{a list of priors for model parameters.}

\item{tuning}{a list of tuning parameters  for model parameters.}

\item{starting}{a list of starting values for model parameters.}

\item{mcmc_param}{a list of MCMC parameters that contains, 
niter: the number of iterations;
nburn: the number of burn-in samples;
nthin: the number of steps to obtain thinned samples.}
}
\description{
This function fits a regression model of which the error term is an 
MTD model with a pre-specified stationary marginal distribution.
}
\references{
Zheng, X., Kottas, A., and Sansó, B. (2021),
"On construction and estimation of stationary mixture transition distribution models,"
\emph{arXiv:2010.12696}.
}
